♻️ Refactor component structure and update imports in Editor and UploadProgramButton
